requirements.txt
torch>=2.0.0
librosa>=0.9.2
numpy>=1.22.0
moviepy>=1.0.3
openai-whisper>=20230314
transformers>=4.28.0
tortoise-tts>=2.4.2
opencv-python>=4.7.0
pydub>=0.25.1
soundfile>=0.12.1
demucs>=4.0.0 